Found a better alpha for NaiveBayes algorithm
I tried a couple different alpha (actually a tried many diffrent small alphas but after trying on and on and getting the same result I started to try aggresively a better alpha jumping to highers and highers values). After the specific alpha that I'm suggesting I notice that the accuracy test (using the test data not the tranning data again) reach 0.6470588235294118 and then it doesn't matter if I change the alpha anymore that it remains the same. I hope it helps this helps out a bit as the accuracy with the test data is 0.5228758169934641.